Split suspension registry by platform to avoid link errors on SJS
